Enemy base class setup
just need to override the stat methods in subclasses of enemy.

next figure out a way to store enemy constructors in a collection so we can pick and choose what enemies are in a level. then make enemymanager randomly spawn enemies of a random amount in the handlespawning().

also make a way to convert elapsed time into a level range for the enemies.