just needed to delay adding renderable to renderlayers_
enemies can be made, everything works.

projectile pool (and gameobject pools) can now be set as expandable when constructed, allowing more to be added to the pool if needed.

this does function as intended until around 10000 projectiles, where it crashes. ideally i fix that but what if we just don't reach 10000 projectiles.

next figure out a level/time formula and implement it. make enemy spawns also base their level off of that value.