Enemy sprite atlas progress
rn i need to find out how to store a SpriteAnimation per enemy base class and have it created, without having to overload loadinfofromjson in every single enemy class. we could then also store the texturestartpos, and size there too.

other than that, if the reference to the rect, textStartPos, and size stays alive, it is done and functional.